split sync and async nodes
