+++ src/node/mod.rs
@@ -1,6 +1,9 @@
-#[allow(clippy::module_inception)]
-mod node;
+#[cfg(feature = "async")]
+mod r#async;
 mod node_path;
+mod sync;
 
-pub(crate) use node::{PyArrayOrGroup, PyAsyncArrayOrGroup, PyAsyncNode, PyNode};
+#[cfg(feature = "async")]
+pub(crate) use r#async::{PyAsyncArrayOrGroup, PyAsyncNode};
 pub(crate) use node_path::PyNodePath;
+pub(crate) use sync::{PyArrayOrGroup, PyNode};
